+++ native/cobbles.c
@@ -3,6 +3,7 @@
 #include <iconv.h>
 #include <stdbool.h>
 #include <errno.h>
+#include <hb.h>
 
 Cobbles * FUNC_NAME(cobbles_init)(CobblesEncoding encoding) {
     Cobbles * cobbles = calloc(1, sizeof(Cobbles));
@@ -15,6 +16,7 @@ Cobbles * FUNC_NAME(cobbles_init)(CobblesEncoding encoding) {
 
     cobbles->error_code = error;
     cobbles->encoding = encoding;
+    cobbles->hb_unicode_funcs = hb_unicode_funcs_get_default();
 
     return cobbles;
 }
@@ -32,6 +34,34 @@ void FUNC_NAME(cobbles_destroy)(Cobbles * cobbles) {
 int FUNC_NAME(cobbles_get_error)(Cobbles * cobbles) {
     return cobbles->error_code;
 }
+
+int FUNC_NAME(cobbles_guess_string_script)(Cobbles * cobbles, char * text) {
+    hb_script_t script = HB_SCRIPT_UNKNOWN;
+    hb_direction_t direction = HB_DIRECTION_LTR;
+    char script_str[4];
+    const uint32_t * code_points = _cobbles_get_code_points(cobbles, text);
+
+    for (size_t index = 0; code_points[index] != 0; index++) {
+        script = hb_unicode_script(
+            cobbles->hb_unicode_funcs, code_points[index]);
+
+        if (script != HB_SCRIPT_COMMON &&
+        script != HB_SCRIPT_INHERITED &&
+        script != HB_SCRIPT_UNKNOWN) {
+            break;
+        }
+    }
+
+    direction = hb_script_get_horizontal_direction(script);
+    hb_tag_to_string(hb_script_to_iso15924_tag(script), script_str);
+
+    return (script_str[0] << 24) |
+        (script_str[1] << 16) |
+        (script_str[2] << 8) |
+        script_str[3] |
+        (direction == HB_DIRECTION_RTL ? 1 << 31 : 0);
+}
+
 
 int _cobbles_bytes_read_int(uint8_t * bytes, int index) {
     return bytes[index] |
@@ -94,7 +124,7 @@ const char* _cobbles_encode_string(Cobbles * cobbles,
     const size_t inputLength = _cobbles_string_length(cobbles, input);
     const size_t bufferSize = 128;
     char buffer[bufferSize];
-    char * bufferCursor = &buffer;
+    char * bufferCursor = (char *) &buffer;
     const char * inputCursor = input;
     size_t inputBytesLeft = inputLength;
     size_t bufferBytesLeft = bufferSize;
@@ -109,10 +139,10 @@ const char* _cobbles_encode_string(Cobbles * cobbles,
 
     while (inputBytesLeft > 0) {
         bufferBytesLeft = bufferSize;
-        bufferCursor = &buffer;
+        bufferCursor = (char *) &buffer;
 
         size_t result = iconv(converter,
-            &inputCursor, &inputBytesLeft,
+            (char **) &inputCursor, &inputBytesLeft,
             &bufferCursor, &bufferBytesLeft);
 
         if (result == (size_t)(-1)) {
@@ -161,8 +191,25 @@ const char* _cobbles_get_utf8_string(Cobbles * cobbles, const char * input) {
         input);
 }
 
+const uint32_t* _cobbles_get_code_points(Cobbles * cobbles, const char * input) {
+    const char * input_encoding;
+
+    switch (cobbles->encoding) {
+        case COBBLES_UTF16: input_encoding = "UTF-16LE"; break;
+        case COBBLES_UTF32: input_encoding = "UTF-32LE"; break;
+        default: input_encoding = "UTF-8"; break;
+    }
+
+    return (const uint32_t*) _cobbles_encode_string(cobbles,
+        input_encoding,
+        "UTF-32LE",
+        input);
+}
+
+
 #ifdef LIBHL_EXPORTS
 DEFINE_PRIM(_ABSTRACT(Cobbles), cobbles_init, _I32);
 DEFINE_PRIM(_VOID, cobbles_destroy, _ABSTRACT(Cobbles));
 DEFINE_PRIM(_I32, cobbles_get_error, _ABSTRACT(Cobbles));
+DEFINE_PRIM(_I32, cobbles_guess_string_script, _ABSTRACT(Cobbles) _BYTES);
 #endif